apps/api/dist-intel/requirements-bundle.txt
torch @ https://huggingface.co/datasets/totoku/universal-macos-torch/resolve/main/torch-2.11.0a0%2Bgit51e04aa-cp311-cp311-macosx_15_0_universal2.whl
torchvision @ https://huggingface.co/datasets/totoku/universal-macos-torch/resolve/main/torchvision-0.17.2%2Bc1d70fe-cp311-cp311-macosx_10_9_universal2.whl
torchaudio @ https://huggingface.co/datasets/totoku/universal-macos-torch/resolve/main/torchaudio-2.11.0a0%2Be123269-cp311-cp311-macosx_10_9_universal2.whl
accelerate==1.9
addict==2.4.0
albumentations==2.0.8
anyio==4.9.0
av==16.1.0
azure-storage-blob>=12.0.0
azureml==0.2.7
boltons==23.0.0
boto3==1.35.53
botocore==1.35.99
einops==0.8.1
fastapi==0.128.0
diffusers @ git+https://github.com/huggingface/diffusers.git
ffmpeg==1.4
ftfy==6.3.1
fvcore==0.1.5.post20221221
gdown==5.2.0
gguf==0.17.1
huggingface_hub==0.34.0
imageio==2.37.0
iopath==0.1.10
ipdb==0.13.13
jsonschema==4.25.1
librosa==0.11.0
loguru==0.7.3
mano==0.6.0
zstandard==0.25.0
matplotlib==3.10.8
more_itertools==10.8.0
moviepy==1.0.3
numpy==1.26.4
numba==0.62.1; sys_platform == "darwin"
llvmlite==0.45.1; sys_platform == "darwin"
omegaconf==2.3.0
opencv_python_headless==4.11.0.86
packaging==25.0
peft==0.17.1
Pillow<=12.0.0
plyfile==1.1.3
python-multipart==0.0.20
pooch==1.8.2
prettytable==3.17.0
protobuf==4.25.8
psutil==7.2.1
pydantic==2.12.5
pydash==8.0.5
pyloudnorm==0.1.1
python-dotenv==1.2.1
ray @ https://s3-us-west-2.amazonaws.com/ray-wheels/latest/ray-3.0.0.dev0-cp312-cp312-win_amd64.whl; sys_platform == "win32" and python_version == "3.12"
ray==2.40.0; sys_platform != "win32"
regex==2024.9.11
Requests==2.32.5
rotary_embedding_torch==0.8.9
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.0
scikit-image>=0.22.0
soundfile==0.13.1
sympy==1.14.0
tenacity==8.2.3
tensorboardX==2.6.4
termcolor==3.3.0
tifffile==2023.4.12
transformers==4.57.1
timm==1.0.22
tqdm==4.67.1
trimesh==4.10.1
turbojpeg==0.0.2
typer==0.21.0
typing_extensions==4.15.0
ujson==5.10.0
uvicorn==0.35.0
urllib3==2.6.2
webcolors==25.10.0
websocket_client==1.8.0
wsproto==1.2.0
xfuser==0.4.5; sys_platform != "darwin" or platform_machine != "x86_64"
yacs==0.1.8
yapf==0.40.2
pillow-avif-plugin==1.5.3
facexlib==0.3.0
insightface @ https://raw.githubusercontent.com/cobanov/insightface_windows/d97fe110e1e5653f9814033291524147f2111d0f/whls/insightface-0.7.3-cp312-cp312-win_amd64.whl ; sys_platform == "win32"
mediapipe==0.10.21
mmengine==0.10.7
ptlflow
raft
Cython>=0.29.36
rtree
git+https://github.com/xinyu1205/recognize-anything.git
git+https://github.com/totokunda/sam2.git ; sys_platform != "darwin" or platform_machine != "x86_64"
git+https://github.com/totokunda/simplepyutils.git
git+https://github.com/totokunda/florch.git
git+https://github.com/isarandi/smplfitter.git
fairscale
pycocotools
rembg[cpu]
onnx==1.20.0
onnxruntime==1.20.1